new update on the user-Dashboard, SquareCard and  user-RequestForm
fetch datas
