+++ src/buckup/SquareCard.jsx
@@ -1,12 +1,14 @@
 import React from "react";
 import { useNavigate } from "react-router-dom"; // Import useNavigate
 import "./SquareCard.css";
+import PropTypes from "prop-types"; // Import prop-types
 
-const SquareCard = ({ title, deviceId, dateLost, imageSrc }) => {
+const SquareCard = ({ title = "Unknown Item", deviceId = "N/A", dateLost = "Not Available", imageSrc, category = "Miscellaneous" }) => {
+
     const navigate = useNavigate(); // Initialize navigate function
 
     const handleRetrieveClick = () => {
-        navigate("/urf"); // Change "/next-page" to your target route
+        navigate("/request-form", { state: { title, dateLost, category } });
     };
 
     return (
@@ -17,10 +19,24 @@ const SquareCard = ({ title, deviceId, dateLost, imageSrc }) => {
                 <button className="square-card-button" onClick={handleRetrieveClick}>
                     Retrieve
                 </button>
-                <img src={imageSrc} alt="Device Icon" className="square-card-icon" />
+                <img 
+                    src={imageSrc || "/default-image.png"} // Provide a fallback image
+                    alt="Device Icon" 
+                    className="square-card-icon" 
+                />
             </div>
         </div>
     );
 };
 
+// ✅ Add PropTypes to catch type errors
+SquareCard.propTypes = {
+    title: PropTypes.string.isRequired,
+    deviceId: PropTypes.string.isRequired, // Ensure deviceId is always a string
+    dateLost: PropTypes.string.isRequired, // Ensure dateLost is a string (date in YYYY-MM-DD format)
+    imageSrc: PropTypes.string.isRequired, // Ensure imageSrc is a valid string
+    category: PropTypes.string.isRequired,
+};
+
 export default SquareCard;
+
